+++ godot-client/tests/golden/vprocenum/tables/vprocenum_kind_row_table.gd
@@ -18,9 +18,7 @@ func _init(p_local_db: LocalDatabase) -> void:
 	count_ = VprocenumKindRowCountBTreeIndex.new(p_local_db)
 	resource_name_ = VprocenumKindRowResourceNameBTreeIndex.new(p_local_db)
 	script_ = VprocenumKindRowScriptBTreeIndex.new(p_local_db)
-	on_insert(_emit_inserted)
-	on_update(_emit_updated)
-	on_delete(_emit_deleted)
+	_relay_row_signals()
 
 func iter() -> Array[VprocenumKindRow]:
 	var rows: Array[_ModuleTableType] = super()
@@ -79,12 +77,3 @@ func find_by_create_(value: int) -> Array[VprocenumKindRow]:
 
 func first_by_create_(value: int) -> VprocenumKindRow:
 	return first_by(&"create_", value)
-
-func _emit_inserted(row: _ModuleTableType) -> void:
-	inserted.emit(row as VprocenumKindRow)
-
-func _emit_updated(old_row: _ModuleTableType, new_row: _ModuleTableType) -> void:
-	updated.emit(old_row as VprocenumKindRow, new_row as VprocenumKindRow)
-
-func _emit_deleted(row: _ModuleTableType) -> void:
-	deleted.emit(row as VprocenumKindRow)