+++ godot-client/tests/golden/vsubmod/tables/vsubmod_auth_auth_session_table.gd
@@ -12,9 +12,7 @@ func _init(p_local_db: LocalDatabase) -> void:
 	super(p_local_db)
 	_table_name = &"auth.auth_session"
 	id = VsubmodAuthAuthSessionIdUniqueIndex.new(p_local_db)
-	on_insert(_emit_inserted)
-	on_update(_emit_updated)
-	on_delete(_emit_deleted)
+	_relay_row_signals()
 
 func iter() -> Array[VsubmodAuthAuthSession]:
 	var rows: Array[_ModuleTableType] = super()
@@ -55,12 +53,3 @@ func find_by_token(value: String) -> Array[VsubmodAuthAuthSession]:
 
 func first_by_token(value: String) -> VsubmodAuthAuthSession:
 	return first_by(&"token", value)
-
-func _emit_inserted(row: _ModuleTableType) -> void:
-	inserted.emit(row as VsubmodAuthAuthSession)
-
-func _emit_updated(old_row: _ModuleTableType, new_row: _ModuleTableType) -> void:
-	updated.emit(old_row as VsubmodAuthAuthSession, new_row as VsubmodAuthAuthSession)
-
-func _emit_deleted(row: _ModuleTableType) -> void:
-	deleted.emit(row as VsubmodAuthAuthSession)