+++ godot-client/tests/golden/vsum/tables/vsum_note_table.gd
@@ -10,9 +10,7 @@ signal deleted(row: VsumNote)
 func _init(p_local_db: LocalDatabase) -> void:
 	super(p_local_db)
 	_table_name = &"note"
-	on_insert(_emit_inserted)
-	on_update(_emit_updated)
-	on_delete(_emit_deleted)
+	_relay_row_signals()
 
 func iter() -> Array[VsumNote]:
 	var rows: Array[_ModuleTableType] = super()
@@ -43,12 +41,3 @@ func find_by_text(value: String) -> Array[VsumNote]:
 
 func first_by_text(value: String) -> VsumNote:
 	return first_by(&"text", value)
-
-func _emit_inserted(row: _ModuleTableType) -> void:
-	inserted.emit(row as VsumNote)
-
-func _emit_updated(old_row: _ModuleTableType, new_row: _ModuleTableType) -> void:
-	updated.emit(old_row as VsumNote, new_row as VsumNote)
-
-func _emit_deleted(row: _ModuleTableType) -> void:
-	deleted.emit(row as VsumNote)