+++ godot-client/tests/golden/vtypes/tables/vtypes_one_uuid_table.gd
@@ -10,9 +10,7 @@ signal deleted(row: VtypesOneUuid)
 func _init(p_local_db: LocalDatabase) -> void:
 	super(p_local_db)
 	_table_name = &"one_uuid"
-	on_insert(_emit_inserted)
-	on_update(_emit_updated)
-	on_delete(_emit_deleted)
+	_relay_row_signals()
 
 func iter() -> Array[VtypesOneUuid]:
 	var rows: Array[_ModuleTableType] = super()
@@ -43,12 +41,3 @@ func find_by_u(value: PackedByteArray) -> Array[VtypesOneUuid]:
 
 func first_by_u(value: PackedByteArray) -> VtypesOneUuid:
 	return first_by(&"u", value)
-
-func _emit_inserted(row: _ModuleTableType) -> void:
-	inserted.emit(row as VtypesOneUuid)
-
-func _emit_updated(old_row: _ModuleTableType, new_row: _ModuleTableType) -> void:
-	updated.emit(old_row as VtypesOneUuid, new_row as VtypesOneUuid)
-
-func _emit_deleted(row: _ModuleTableType) -> void:
-	deleted.emit(row as VtypesOneUuid)