+++ godot-client/tests/golden/vtypes2/tables/vtypes_2_event_log_table.gd
@@ -10,9 +10,7 @@ signal deleted(row: Vtypes2EventLog)
 func _init(p_local_db: LocalDatabase) -> void:
 	super(p_local_db)
 	_table_name = &"event_log"
-	on_insert(_emit_inserted)
-	on_update(_emit_updated)
-	on_delete(_emit_deleted)
+	_relay_row_signals()
 
 func iter() -> Array[Vtypes2EventLog]:
 	var rows: Array[_ModuleTableType] = super()
@@ -43,12 +41,3 @@ func find_by_msg(value: String) -> Array[Vtypes2EventLog]:
 
 func first_by_msg(value: String) -> Vtypes2EventLog:
 	return first_by(&"msg", value)
-
-func _emit_inserted(row: _ModuleTableType) -> void:
-	inserted.emit(row as Vtypes2EventLog)
-
-func _emit_updated(old_row: _ModuleTableType, new_row: _ModuleTableType) -> void:
-	updated.emit(old_row as Vtypes2EventLog, new_row as Vtypes2EventLog)
-
-func _emit_deleted(row: _ModuleTableType) -> void:
-	deleted.emit(row as Vtypes2EventLog)