+++ godot-client/tests/golden/vtypes2/tables/vtypes_2_seq_row_table.gd
@@ -12,9 +12,7 @@ func _init(p_local_db: LocalDatabase) -> void:
 	super(p_local_db)
 	_table_name = &"seq_row"
 	id = Vtypes2SeqRowIdUniqueIndex.new(p_local_db)
-	on_insert(_emit_inserted)
-	on_update(_emit_updated)
-	on_delete(_emit_deleted)
+	_relay_row_signals()
 
 func iter() -> Array[Vtypes2SeqRow]:
 	var rows: Array[_ModuleTableType] = super()
@@ -55,12 +53,3 @@ func find_by_v(value: int) -> Array[Vtypes2SeqRow]:
 
 func first_by_v(value: int) -> Vtypes2SeqRow:
 	return first_by(&"v", value)
-
-func _emit_inserted(row: _ModuleTableType) -> void:
-	inserted.emit(row as Vtypes2SeqRow)
-
-func _emit_updated(old_row: _ModuleTableType, new_row: _ModuleTableType) -> void:
-	updated.emit(old_row as Vtypes2SeqRow, new_row as Vtypes2SeqRow)
-
-func _emit_deleted(row: _ModuleTableType) -> void:
-	deleted.emit(row as Vtypes2SeqRow)